refactor: firebase ディレクトリ直下に repository ファイルを配置する
