+++ lib/infrastructure/firebase/user_repository.dart
@@ -4,8 +4,8 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:mood_trend_flutter/domain/app_user.dart';
 
-import '../../../domain/app_exception.dart';
-import '../firebase_provider.dart';
+import '../../domain/app_exception.dart';
+import 'firebase_provider.dart';
 
 /// [FirebaseUserRepository] のインスタンスを提供する [Provider]
 final firebaseUserRepositoryProvider = Provider<FirebaseUserRepository>(
