+++ lib/presentation/root_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:mood_trend_flutter/infrastructure/firebase/app_user/user_repository.dart';
+import 'package:mood_trend_flutter/infrastructure/firebase/user_repository.dart';
 import 'package:mood_trend_flutter/presentation/components/loading.dart';
 import 'package:mood_trend_flutter/presentation/home_page.dart';
 
-import '../infrastructure/firebase/authentication/auth_repository.dart';
+import '../infrastructure/firebase/auth_repository.dart';
 
 final rootPageKey = Provider((ref) => GlobalKey<NavigatorState>());
